+++ browser-extension/src/core/balance/current-assets.ts
@@ -4,12 +4,14 @@ import { balancesStore } from '@src/prun-api/data/balances';
 import { cxosStore } from '@src/prun-api/data/cxos';
 import { fxosStore } from '@src/prun-api/data/fxos';
 import {
+  selfConditions,
   partnerCurrentConditions,
   sumLoanInstallments,
   sumAccountsPayable,
-  sumMaterialsPayable,
+  sumDeliveries,
+  sumPendingMaterialsPickup,
+  sumMaterialsShipment,
   sumMaterialsPickup,
-  selfConditions,
 } from '@src/core/balance/contract-conditions';
 import { getPrice } from '@src/fio/cx';
 import { sumMapValues } from '@src/core/balance/utils';
@@ -76,12 +78,16 @@ const marketListedMaterials = computed(() => {
   return sumBy(sellOrders, x => getPrice(x.material.ticker) * x.amount);
 });
 
-// After DELIVERY/PROVISION was fulfilled by partner,
-// the player must pick up the materials via COMEX_PURCHASE_PICKUP condition.
-// We'll consider these materials as a part of current assets, regardless of condition
-// deadline, since they can be picked up at any time.
 const materialsToReceive = computed(
-  () => sumMaterialsPayable(partnerCurrentConditions) + sumMaterialsPickup(selfConditions),
+  () =>
+    sumDeliveries(partnerCurrentConditions) +
+    sumMaterialsShipment(partnerCurrentConditions) +
+    sumMaterialsPickup(selfConditions) +
+    // After all dependencies are fulfilled,
+    // the player must pick up the materials via COMEX_PURCHASE_PICKUP condition.
+    // We'll consider these materials as a part of current assets, regardless of
+    // COMEX_PURCHASE_PICKUP deadline, since they can be picked up at any time.
+    sumPendingMaterialsPickup(selfConditions),
 );
 
 const total = computed(() => {